Policy Authorization in Controller (10.8)

+++ app/Http/Controllers/JobController.php
@@ -7,9 +7,12 @@ use Illuminate\View\View;
 use Illuminate\Http\RedirectResponse;
 use App\Models\Job;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     // @desc Show all job listings
     // @route GET /jobs
     public function index(): View
@@ -83,6 +86,9 @@ class JobController extends Controller
     // @route GET /jobs/{$id}/edit
     public function edit(Job $job): View
     {
+        // Check if user is authorized (owns the job listing)
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -90,6 +96,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {
+        // Check if user is authorized (owns the job listing)
+        $this->authorize('update', $job);
+
         // Convert "true"/"false" to actual boolean values
         $request->merge([
             'remote' => filter_var($request->input('remote'), FILTER_VALIDATE_BOOLEAN),
@@ -138,6 +147,9 @@ class JobController extends Controller
     // @route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
+        // Check if user is authorized (owns the job listing)
+        $this->authorize('delete', $job);
+
         // If logo, then delete it
         if ($job->company_logo) {
             Storage::delete('public/logos/' . $job->company_logo);
